Shuffle around some `#[cfg]`'d code
